pyproject.toml
[project]
name = "codereview-env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "huggingface-hub>=1.8.0",
    "openai>=1.14.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.29.0",
]
